+++ web/assets/js/app.js
@@ -2,37 +2,7 @@ const $ = require('jquery');
 
 require('bootstrap-sass');
 
-var TexthubApp = {
-  initialize: function($wrapper) {
-    this.$wrapper = $wrapper;
-
-    this.$wrapper
-      .find('#new-text-modal')
-      .on(
-        'shown.bs.modal',
-        this.handleNewTextModalOpen
-      );
-
-    this.$wrapper
-      .find('#edit-text-title-modal')
-      .on(
-        'shown.bs.modal',
-        this.handleEditTextTitleModalOpen
-      );
-  },
-
-  handleEditTextTitleModalOpen: function() {
-    $(this)
-      .find('#appbundle_text_title')
-      .select();
-  },
-
-  handleNewTextModalOpen: function() {
-    $(this)
-      .find('#appbundle_text_new_title')
-      .select();
-  }
-};
+const TexthubApp = require('./TexthubApp');
 
 $(document).ready(function() {
   var $body = $('body');
